Refactor k-box
