+++ panel/src/ui/components/Box.vue
@@ -1,5 +1,9 @@
 <template>
-  <div :data-theme="theme" class="k-box" v-on="$listeners">
+  <div
+    :data-theme="theme"
+    class="k-box text-sm"
+    v-on="$listeners"
+  >
     <slot>
       <k-text v-html="text" />
     </slot>
@@ -9,7 +13,13 @@
 <script>
 export default {
   props: {
+    /**
+     * `positive`|`negative`|`info`|`none`
+     */
     theme: String,
+    /**
+     * Text to display in the box
+     */
     text: String
   }
 };
@@ -18,14 +28,13 @@ export default {
 <style lang="scss">
 .k-box {
   word-wrap: break-word;
-  font-size: $text-sm;
 }
 .k-box:not([data-theme="none"]) {
-  background: lighten($color-light-grey, 25%);
-  border-radius: $rounded-xs;
+  background: lighten($color-gray-500, 25%);
+  border-radius: $rounded-sm;
   padding: 0.375rem 0.75rem;
   line-height: 1.25rem;
-  border-left: 2px solid $color-light-grey;
+  border-left: 2px solid $color-gray-500;
   padding: 0.5rem 1.5rem;
 }
 .k-box[data-theme="code"] {
@@ -83,13 +92,13 @@ export default {
   align-items: center;
   flex-direction: column;
   background: $color-background;
-  border-radius: $rounded-xs;
+  border-radius: $rounded-sm;
   color: $color-gray-700;
   border: 1px dashed $color-border;
 }
 .k-box[data-theme="empty"] .k-icon {
   margin-bottom: 0.5rem;
-  color: $color-light-grey;
+  color: $color-gray-500;
 }
 .k-box[data-theme="empty"] p {
   color: $color-gray-700;
